+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/statements/ParserAbstract.java
@@ -222,7 +222,7 @@ public abstract class ParserAbstract<S extends AbstractDatabase> {
         checkLocation(child, QNameParser.getFirstNameCtx(ids).getStart());
     }
 
-    private void checkLocation(PgStatement statement, Token errToken) {
+    protected void checkLocation(PgStatement statement, Token errToken) {
         if (isRefMode() || fileName == null) {
             return;
         }